More refactoring for organization. Removed unused StudentSearchController.java
