+++ src/main/java/app/FirebaseAppBuilder.java
@@ -43,18 +43,18 @@ import interface_adapter.student_explore_clubs.ExploreClubsViewModel;
 import interface_adapter.student_home.StudentHomeController;
 import interface_adapter.student_home.StudentHomePresenter;
 import interface_adapter.student_home.StudentHomeViewModel;
-import interface_adapter.student_home.dislike.StudentDislikeController;
-import interface_adapter.student_home.dislike.StudentDislikePresenter;
-import interface_adapter.student_home.join_club.JoinClubController;
-import interface_adapter.student_home.join_club.JoinClubPresenter;
-import interface_adapter.student_home.leave_club.LeaveClubController;
-import interface_adapter.student_home.leave_club.LeaveClubPresenter;
-import interface_adapter.student_home.like.StudentLikeController;
-import interface_adapter.student_home.like.StudentLikePresenter;
-import interface_adapter.student_home.show_clubs.StudentShowClubsController;
-import interface_adapter.student_home.show_clubs.StudentShowClubsPresenter;
-import interface_adapter.student_home.show_posts.StudentShowPostsController;
-import interface_adapter.student_home.show_posts.StudentShowPostsPresenter;
+import interface_adapter.student_home.student_dislike.StudentDislikeController;
+import interface_adapter.student_home.student_dislike.StudentDislikePresenter;
+import interface_adapter.student_home.student_join_club.JoinClubController;
+import interface_adapter.student_home.student_join_club.JoinClubPresenter;
+import interface_adapter.student_home.student_leave_club.LeaveClubController;
+import interface_adapter.student_home.student_leave_club.LeaveClubPresenter;
+import interface_adapter.student_home.student_like.StudentLikeController;
+import interface_adapter.student_home.student_like.StudentLikePresenter;
+import interface_adapter.student_home.student_show_clubs.StudentShowClubsController;
+import interface_adapter.student_home.student_show_clubs.StudentShowClubsPresenter;
+import interface_adapter.student_home.student_show_posts.StudentShowPostsController;
+import interface_adapter.student_home.student_show_posts.StudentShowPostsPresenter;
 import interface_adapter.student_profile.StudentProfileController;
 import interface_adapter.student_profile.StudentProfilePresenter;
 import interface_adapter.student_profile.StudentProfileViewModel;
