--- src/main/java/interface_adapter/student_home/search/StudentSearchController.java
@@ -1,8 +1,0 @@
-package interface_adapter.student_home.search;
-
-/**
- * The Controller for the search clubs usecase on the student home page.
- */
-public class StudentSearchController {
-    // Left for David.
-}
