+++ src/main/java/interface_adapter/student_home/student_dislike/StudentDislikePresenter.java
@@ -1,4 +1,4 @@
-package interface_adapter.student_home.like;
+package interface_adapter.student_home.student_dislike;
 
 import java.util.List;
 import java.util.Map;
@@ -6,32 +6,32 @@ import java.util.Map;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.student_home.StudentHomeState;
 import interface_adapter.student_home.StudentHomeViewModel;
-import use_case.student_like.StudentLikeOutputBoundary;
-import use_case.student_like.StudentLikeOutputData;
+import use_case.student_dislike.StudentDislikeOutputBoundary;
+import use_case.student_dislike.StudentDislikeOutputData;
 
 /**
- * The presenter for the like usecase.
+ * The presenter for the dislike usecase.
  */
-public class StudentLikePresenter implements StudentLikeOutputBoundary {
+public class StudentDislikePresenter implements StudentDislikeOutputBoundary {
     private final StudentHomeViewModel viewModel;
     private final ViewManagerModel viewManagerModel;
 
-    public StudentLikePresenter(StudentHomeViewModel viewModel, ViewManagerModel viewManagerModel) {
+    public StudentDislikePresenter(StudentHomeViewModel viewModel, ViewManagerModel viewManagerModel) {
         this.viewModel = viewModel;
         this.viewManagerModel = viewManagerModel;
     }
 
     @Override
-    public void prepareSuccessView(StudentLikeOutputData data) {
+    public void prepareSuccessView(StudentDislikeOutputData data) {
         final Map<String, Object> postData = data.getPostData();
 
         final StudentHomeState currentState = viewModel.getState();
 
-        // the old post data, which as to be updated to reflect the new like.
+        // the old post data, which as to be updated to reflect the new dislike.
         final Map<String, List<Map<String, Object>>> postsByClub = currentState.getPostData();
         final int indexToUpdate;
         for (Map<String, Object> post : postsByClub.get(data.getClubName())) {
-            if (post.get("time") == postData.get("time") && post.get("date") == postData.get("date")) {
+            if (post.get("time") == postData.get("time") && post.get("state") == postData.get("state")) {
                 indexToUpdate = postsByClub.get(data.getClubName()).indexOf(post);
                 postsByClub.get(data.getClubName()).set(indexToUpdate, postData);
                 break;
@@ -39,10 +39,10 @@ public class StudentLikePresenter implements StudentLikeOutputBoundary {
         }
         currentState.setPostData(postsByClub);
         viewModel.setState(currentState);
-        viewModel.firePropertyChanged("liked");
+        viewModel.firePropertyChanged("disliked");
 
         viewManagerModel.setState(viewModel.getViewName());
-        viewManagerModel.firePropertyChanged("liked");
+        viewManagerModel.firePropertyChanged("disliked");
     }
 
     @Override
